@@ -44,94 +40,96 @@
   },
   "homepage": "https://github.com/andela-oaboluwarin/document-management-system#readme",
   "devDependencies": {
-    "chai": "^3.5.0",
-    "compression": "1.6.1",
-    "cross-env": "1.0.7",
-    "css-loader": "0.27.3",
+    "@babel/core": "^7.7.4",
+    "@babel/node": "^7.7.4",
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
+    "@babel/preset-env": "^7.7.4",
+    "@babel/preset-react": "^7.7.4",
+    "babel-loader": "^8.0.6",
+    "chai": "^4.2.0",
+    "compression": "1.7.4",
+    "cross-env": "6.0.3",
+    "css-loader": "3.2.1",
     "env2": "^2.1.1",
-    "enzyme": "^2.2.0",
-    "eslint": "^3.18.0",
-    "eslint-config-airbnb-base": "^11.1.0",
-    "eslint-import-resolver-webpack": "^0.8.1",
+    "enzyme": "^3.10.0",
+    "eslint": "^6.7.2",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-import-resolver-webpack": "^0.11.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "5.0.1",
-    "eslint-watch": "^3.0.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "7.17.0",
+    "eslint-watch": "^6.0.1",
     "eventsource-polyfill": "0.9.6",
-    "expect": "1.19.0",
-    "extract-text-webpack-plugin": "1.0.1",
+    "expect": "24.9.0",
+    "extract-text-webpack-plugin": "3.0.2",
     "faker": "^4.1.0",
-    "file-loader": "0.8.5",
-    "html-webpack-plugin": "^2.28.0",
+    "file-loader": "5.0.2",
+    "html-webpack-plugin": "^3.2.0",
     "istanbul": "^0.4.5",
-    "jsdom": "8.5.0",
-    "mocha": "^2.4.5",
+    "jsdom": "15.2.1",
+    "mocha": "^6.2.2",
     "mocha-lcov-reporter": "^1.3.0",
-    "nock": "8.0.0",
+    "nock": "11.7.0",
     "react-addons-test-utils": "^15.0.2",
-    "redux-immutable-state-invariant": "1.2.3",
-    "redux-mock-store": "1.0.2",
-    "rimraf": "2.5.2",
+    "redux-immutable-state-invariant": "2.1.0",
+    "redux-mock-store": "1.5.3",
+    "rimraf": "3.0.0",
     "selenium-download": "^2.0.9",
     "selenium-standalone": "^6.2.0",
-    "sinon": "^2.1.0",
-    "style-loader": "0.13.1",
-    "supertest": "^3.0.0",
-    "url-loader": "0.5.7",
-    "webpack-dev-middleware": "1.6.1",
-    "webpack-hot-middleware": "2.10.0"
+    "sinon": "^7.5.0",
+    "style-loader": "1.0.1",
+    "supertest": "^4.0.2",
+    "url-loader": "3.0.0",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-middleware": "3.7.2",
+    "webpack-hot-middleware": "2.25.0"
   },
   "dependencies": {
-    "axios": "0.13.1",
-    "babel-cli": "^6.23.0",
-    "babel-core": "^6.24.0",
-    "babel-eslint": "^7.2.1",
+    "@material-ui/core": "^4.7.1",
+    "@material-ui/icons": "^4.5.1",
+    "axios": "0.19.0",
+    "babel-eslint": "^10.0.3",
     "babel-istanbul": "^0.12.2",
-    "babel-loader": "^6.4.1",
     "babel-plugin-react-display-name": "^2.0.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-polyfill": "6.8.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-2": "^6.22.0",
+    "babel-polyfill": "6.26.0",
     "babel-register": "^6.23.0",
     "bcrypt-nodejs": "0.0.3",
-    "cheerio": "^0.22.0",
+    "cheerio": "^1.0.0-rc.3",
     "classnames": "^2.2.5",
     "colors": "^1.1.2",
-    "dotenv": "^4.0.0",
+    "dotenv": "^8.2.0",
     "express": "^4.15.2",
     "font-awesome": "^4.7.0",
-    "froala-editor": "^2.5.1",
-    "jquery": "2.2.3",
-    "jsonwebtoken": "^7.3.0",
+    "froala-editor": "^3.0.6",
+    "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.4",
-    "material-icons": "^0.1.0",
-    "materialize": "^1.0.0",
-    "materialize-css": "^0.98.1",
+    "material-icons": "^0.3.1",
+    "mini-css-extract-plugin": "^0.8.0",
     "morgan": "^1.8.1",
-    "nightwatch": "^0.9.14",
-    "npm-run-all": "1.8.0",
-    "open": "0.0.5",
-    "pg": "^6.1.4",
+    "nightwatch": "^1.3.1",
+    "node-sass": "^4.13.0",
+    "npm-run-all": "4.1.5",
+    "open": "7.0.0",
+    "pg": "^7.14.0",
     "pg-hstore": "^2.3.2",
     "prop-types": "^15.6.2",
-    "react": "^16.7.0-alpha.2",
-    "react-dom": "^16.7.0-alpha.2",
-    "react-froala-wysiwyg": "^2.5.1",
-    "react-redux": "^5.1.1",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
+    "react-froala-wysiwyg": "^3.0.6",
+    "react-redux": "^7.1.3",
     "react-redux-sweetalert": "^1.0.2",
-    "react-router-dom": "^4.3.1",
-    "redux": "^4.0.1",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.4",
     "redux-devtools-extension": "^2.13.5",
     "redux-thunk": "^2.3.0",
-    "sequelize": "^3.30.2",
-    "sequelize-cli": "^2.5.1",
+    "sass-loader": "^8.0.0",
+    "sequelize": "^5.21.2",
+    "sequelize-cli": "^5.5.1",
     "shortid": "^2.2.8",
-    "toastr": "2.1.2",
-    "validator": "^7.0.0",
-    "webpack": "^1.14.0"
+    "toastr": "2.1.4",
+    "validator": "^12.1.0",
+    "webpack": "^4.41.2"
   }
 }
