@@ -8,24 +8,24 @@ import * as types from './actionTypes';
  * @returns  {object} user
  */
 export function retrieveUsersSuccess(users) {
-  return {
-    type: types.RETRIEVE_USERS_SUCCESS,
-    users
-  };
+    return {
+        type: types.RETRIEVE_USERS_SUCCESS,
+        users
+    };
 }
 
 export function updateUserSuccess(user) {
-  return {
-    type: types.UPDATE_USER_SUCCESS,
-    user
-  };
+    return {
+        type: types.UPDATE_USER_SUCCESS,
+        user
+    };
 }
 
 export function getUserInfoSuccess(user) {
-  return {
-    type: types.GET_USER_INFO_SUCCESS,
-    user
-  };
+    return {
+        type: types.GET_USER_INFO_SUCCESS,
+        user
+    };
 }
 
 /**
@@ -34,12 +34,15 @@ export function getUserInfoSuccess(user) {
  * @returns {object}
  */
 export function retrieveUsers() {
-  return dispatch => axios.get('/user')
-    .then((response) => {
-      dispatch(retrieveUsersSuccess(response.data.usersFound));
-    }).catch((error) => {
-      throw (error);
-    });
+    return dispatch =>
+        axios
+            .get('/user')
+            .then(response => {
+                dispatch(retrieveUsersSuccess(response.data.usersFound));
+            })
+            .catch(error => {
+                throw error;
+            });
 }
 
 /**
@@ -49,12 +52,15 @@ export function retrieveUsers() {
  * @returns {object}
  */
 export function deleteUser(id) {
-  return dispatch => axios.delete(`/user/${id}`)
-    .then(() => {
-      dispatch(retrieveUsers());
-    }).catch((error) => {
-      throw (error);
-    });
+    return dispatch =>
+        axios
+            .delete(`/user/${id}`)
+            .then(() => {
+                dispatch(retrieveUsers());
+            })
+            .catch(error => {
+                throw error;
+            });
 }
 
 /**
@@ -64,31 +70,39 @@ export function deleteUser(id) {
  * @returns {object}
  */
 export function updateUserAdmin(user) {
-  return dispatch => axios.put(`/user/${user.id}`, user)
-      .then(() => {
-        dispatch(retrieveUsers());
-      }).catch((error) => {
-        throw (error);
-      });
+    return dispatch =>
+        axios
+            .put(`/user/${user.id}`, user)
+            .then(() => {
+                dispatch(retrieveUsers());
+            })
+            .catch(error => {
+                throw error;
+            });
 }
 
 export function updateUserInfo(userInfo) {
-  const user = JSON.parse(localStorage.getItem('user'));
-  return dispatch => axios.put(`/user/${user.id}`, userInfo)
-    .then((response) => {
-      dispatch(updateUserSuccess(userInfo));
-    }).catch((error) => {
-      throw error;
-    });
+    const user = JSON.parse(localStorage.getItem('user'));
+    return dispatch =>
+        axios
+            .put(`/user/${user.id}`, userInfo)
+            .then(response => {
+                dispatch(updateUserSuccess(userInfo));
+            })
+            .catch(error => {
+                throw error;
+            });
 }
 
 export function getUserInfo() {
-  const user = JSON.parse(localStorage.getItem('user'));
-  return dispatch => axios.get(`/user/${user.id}`)
-    .then((response) => {
-      dispatch(getUserInfoSuccess(response.data));
-    }).catch((error) => {
-      throw error;
-    });
+    const user = JSON.parse(localStorage.getItem('user'));
+    return dispatch =>
+        axios
+            .get(`/user/${user.id}`)
+            .then(response => {
+                dispatch(getUserInfoSuccess(response.data));
+            })
+            .catch(error => {
+                throw error;
+            });
 }
-
